pyproject.toml
[tool.poetry]
name = "spotifyapitraining"
version = "0.1.0"
description = ""
authors = ["Little-Wretch <nazarr914@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
